refactor: rename getProducts service file to a more generic name
