+++ src/services/index.ts
@@ -1,1 +1,1 @@
-export * from './getProducts';
+export * from './fetchData';
